Unicast channel encryption
Added an encryption layer to the network transport
protocol.
