+++ pkg/net/libp2p/libp2p.go
@@ -258,7 +258,7 @@ func discoverAndListen(
 		return nil, err
 	}
 
-	transport, err := newAuthenticatedTransport(
+	transport, err := newEncryptedAuthenticatedTransport(
 		identity.privKey,
 		stakeMonitor,
 	)
